# api-gateway/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Retail Enterprise Integration Platform API",
    "description": "API Gateway for EAI Retail Integration System. Routes requests to microservices and handles authentication.",
    "version": "1.0.0",
    "contact": { "name": "EAI Team" }
  },
  "servers": [{ "url": "http://localhost:3000", "description": "API Gateway" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "SalesEvent": {
        "type": "object",
        "properties": {
          "eventId": { "type": "string" },
          "source": { "type": "string", "enum": ["pos", "ecommerce"] },
          "customerId": { "type": "string" },
          "productId": { "type": "string" },
          "quantity": { "type": "integer" },
          "price": { "type": "number" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "Sale": {
        "type": "object",
        "properties": {
          "invoice_no": { "type": "string" },
          "customer_id": { "type": "string" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "product_id": { "type": "string" },
                "product_name": { "type": "string" },
                "quantity": { "type": "integer" },
                "unit_price": { "type": "number" }
              }
            }
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "stock": { "type": "integer" },
          "price": { "type": "number" },
          "category": { "type": "string" }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "order_no": { "type": "string" },
          "customer_id": { "type": "string" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "product_id": { "type": "string" },
                "quantity": { "type": "integer" },
                "unit_price": { "type": "number" }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "admin" },
                  "password": { "type": "string", "example": "admin123" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Login successful" } }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Current user info" } }
      }
    },
    "/api/pos/sales": {
      "get": {
        "tags": ["POS Service"],
        "summary": "List all sales",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of sales" } }
      },
      "post": {
        "tags": ["POS Service"],
        "summary": "Create sale (produces XML, triggers integration)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Sale" } } } },
        "responses": { "201": { "description": "Sale created" } }
      }
    },
    "/api/inventory/products": {
      "get": {
        "tags": ["Inventory Service"],
        "summary": "List all products",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of products" } }
      },
      "post": {
        "tags": ["Inventory Service"],
        "summary": "Create product",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
        "responses": { "201": { "description": "Product created" } }
      }
    },
    "/api/crm/customers": {
      "get": {
        "tags": ["CRM Service"],
        "summary": "List all customers",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of customers" } }
      },
      "post": {
        "tags": ["CRM Service"],
        "summary": "Create customer",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Customer" } } } },
        "responses": { "201": { "description": "Customer created" } }
      }
    },
    "/api/ecommerce/orders": {
      "get": {
        "tags": ["E-Commerce Service"],
        "summary": "List all orders",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of orders" } }
      },
      "post": {
        "tags": ["E-Commerce Service"],
        "summary": "Create order",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
        "responses": { "201": { "description": "Order created" } }
      }
    },
    "/api/accounting/journals": {
      "get": {
        "tags": ["Accounting Service"],
        "summary": "List all journals",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of journals" } }
      }
    },
    "/api/integration/events": {
      "get": {
        "tags": ["Integration Service"],
        "summary": "List processed events",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of events" } }
      },
      "post": {
        "tags": ["Integration Service"],
        "summary": "Submit integration event",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SalesEvent" } } } },
        "responses": { "201": { "description": "Event processed" } }
      }
    },
    "/api/translator/xml-to-json": {
      "post": {
        "tags": ["Message Translator"],
        "summary": "Transform XML to JSON",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/xml": { "schema": { "type": "string" } } } },
        "responses": { "200": { "description": "Transformed JSON" } }
      }
    },
    "/api/router/route": {
      "post": {
        "tags": ["Message Router"],
        "summary": "Route message based on content",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SalesEvent" } } } },
        "responses": { "200": { "description": "Routing result" } }
      }
    }
  }
}